third_party/WebKit/LayoutTests/platform/win/fast/css/fontfaceset-add-remove-expected.txt

Tests that adding/removing a FontFace to/from FontFaceSet trigger font update.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL testElement.offsetWidth should be 58. Was 57.
FAIL testElement.offsetWidth should be 113. Was 112.
FAIL testElement.offsetWidth should be 58. Was 57.
FAIL testElement.offsetWidth should be 113. Was 112.
FAIL testElement.offsetWidth should be 58. Was 57.
PASS successfullyParsed is true

TEST COMPLETE
abcdefg abcdefg abcdefg
